Less warnings for space.*

+++ src/options_for_QT.h
@@ -14,11 +14,11 @@
 
 #include "uncrustify_types.h"
 
-extern bool QT_SIGNAL_SLOT_found;
-extern int  QT_SIGNAL_SLOT_level;
-extern bool restoreValues;
+extern bool   QT_SIGNAL_SLOT_found;
+extern size_t QT_SIGNAL_SLOT_level;
+extern bool   restoreValues;
 
-void save_set_options_for_QT(int level);
+void save_set_options_for_QT(size_t level);
 void restore_options_for_QT(void);
 
 #endif /* OPTIONS_FOR_QT_H_INCLUDED */
